가게 정보 수정 api 리팩토링
